requirements.txt
streamlit
google-generativeai
python-dateutil
Pillow